Updated config, changed names, moved files.
