+++ app/server/jest.config.js
@@ -1,0 +1,12 @@
+module.exports = {
+    verbose: true,
+    testMatch: [
+        "**/*.test.js"
+    ],
+    collectCoverage: false,
+    coverageDirectory: "./tests/coverage",
+    rootDir: "./",
+    setupFiles: [
+        "dotenv/config"
+    ]
+};
