+++ app/server/package.json
@@ -30,14 +30,5 @@
     "jest": "^28.1.0",
     "randexp": "^0.5.3",
     "supertest": "^6.2.3"
-  },
-  "jest": {
-    "verbose": true,
-    "collectCoverage": false,
-    "coverageDirectory": "./tests/coverage",
-    "rootDir": "./",
-    "setupFiles": [
-      "dotenv/config"
-    ]
   }
 }
