+++ app/server/tests/functions/authFunctions.js
