--- app/server/tests/jest.config.js
@@ -1,7 +1,0 @@
-module.exports = {
-    verbose: true,
-    testMatch: [
-        "**/*.test.js"
-    ],
-    setupFiles: ["dotenv/config"]
-};
